move hnsw package to vector/hnsw
