+++ src/java/org/apache/cassandra/index/sai/disk/vector/CassandraOnDiskHnsw.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.cassandra.index.sai.disk.hnsw;
+package org.apache.cassandra.index.sai.disk.vector;
 
 import java.io.IOException;
 import java.util.Arrays;
@@ -34,6 +34,9 @@ import org.apache.cassandra.index.sai.disk.format.IndexComponent;
 import org.apache.cassandra.index.sai.disk.v1.PerIndexFiles;
 import org.apache.cassandra.index.sai.disk.v1.SegmentMetadata;
 import org.apache.cassandra.index.sai.disk.v1.postings.ReorderingPostingList;
+import org.apache.cassandra.index.sai.disk.vector.hnsw.OnDiskHnswGraph;
+import org.apache.cassandra.index.sai.disk.vector.hnsw.OnDiskVectors;
+import org.apache.cassandra.index.sai.disk.vector.hnsw.VectorCache;
 import org.apache.lucene.index.VectorEncoding;
 import org.apache.lucene.index.VectorSimilarityFunction;
 import org.apache.lucene.util.Bits;
