+++ src/java/org/apache/cassandra/index/sai/disk/vector/ConcurrentVectorValues.java
@@ -16,30 +16,24 @@
  * limitations under the License.
  */
 
-package org.apache.cassandra.index.sai.disk.hnsw;
+package org.apache.cassandra.index.sai.disk.vector;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
-import java.util.ArrayList;
 
-import javax.annotation.concurrent.NotThreadSafe;
-
-import org.apache.cassandra.db.marshal.VectorType;
 import org.apache.cassandra.io.util.SequentialWriter;
 import org.apache.lucene.util.RamUsageEstimator;
 import org.apache.lucene.util.hnsw.RandomAccessVectorValues;
+import org.jctools.maps.NonBlockingHashMapLong;
 
-@NotThreadSafe
-public class CompactionVectorValues implements RamAwareVectorValues
+public class ConcurrentVectorValues implements RamAwareVectorValues
 {
-    private final int dimension;
-    private final ArrayList<ByteBuffer> values = new ArrayList<>();
-    private final VectorType<Float> type;
+    private final int dimensions;
+    private final NonBlockingHashMapLong<float[]> values = new NonBlockingHashMapLong<>();
 
-    public CompactionVectorValues(VectorType<Float> type)
+    public ConcurrentVectorValues(int dimensions)
     {
-        this.dimension = type.dimension;
-        this.type = type;
+        this.dimensions = dimensions;
     }
 
     @Override
@@ -51,28 +45,26 @@ public class CompactionVectorValues implements RamAwareVectorValues
     @Override
     public int dimension()
     {
-        return dimension;
+        return dimensions;
     }
 
     @Override
     public float[] vectorValue(int i)
     {
-        return type.getSerializer().deserializeFloatArray(values.get(i));
+        return values.get(i);
     }
 
     /** return approximate bytes used by the new vector */
-    public long add(int ordinal, ByteBuffer value)
+    public long add(int ordinal, float[] vector)
     {
-        if (ordinal != values.size())
-            throw new IllegalArgumentException(String.format("CVV requires vectors to be added in ordinal order (%d given, expected %d)",
-                                                             ordinal, values.size()));
-        values.add(value);
+        values.put(ordinal, vector);
         return RamEstimation.concurrentHashMapRamUsed(1) + oneVectorBytesUsed();
     }
 
     @Override
     public RandomAccessVectorValues<float[]> copy()
     {
+        // no actual copy required because we always return distinct float[] for distinct vector ordinals
         return this;
     }
 
@@ -81,10 +73,15 @@ public class CompactionVectorValues implements RamAwareVectorValues
         writer.writeInt(size());
         writer.writeInt(dimension());
 
+        // we will re-use this buffer
+        var byteBuffer = ByteBuffer.allocate(dimension() * Float.BYTES);
+        var floatBuffer = byteBuffer.asFloatBuffer();
+
         for (var i = 0; i < size(); i++) {
-            var bb = values.get(i);
-            assert bb != null : "null vector at index " + i + " of " + size();
-            writer.write(bb);
+            floatBuffer.put(vectorValue(i));
+            // bytebuffer and floatBuffer track their positions separately, and we never changed BB's, so don't need to rewind it
+            floatBuffer.rewind();
+            writer.write(byteBuffer);
         }
 
         return writer.position();
@@ -100,6 +97,6 @@ public class CompactionVectorValues implements RamAwareVectorValues
 
     private long oneVectorBytesUsed()
     {
-        return RamUsageEstimator.NUM_BYTES_OBJECT_REF;
+        return Integer.BYTES + Integer.BYTES + (long) dimension() * Float.BYTES;
     }
 }
