+++ src/java/org/apache/cassandra/index/sai/disk/vector/VectorPostings.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.cassandra.index.sai.disk.hnsw;
+package org.apache.cassandra.index.sai.disk.vector;
 
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
